web3py/simplestorage/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
numpy==1.22.0
pandas==1.3.5
parsimonious==0.8.1
protobuf==3.19.1
py-solc-x==1.1.1
pycryptodome==3.12.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rlp==2.0.1
semantic-version==2.8.5
six==1.16.0
toolz==0.11.2
urllib3==1.26.8
varint==1.0.2
web3==5.26.0
websockets==9.1
yarl==1.7.2
